# swagger.yaml
openapi: 3.0.3
swagger: 2.0
info:
  title: vieritage
  version: 1.0.0
  contact: {}
servers:
  - url: http://localhost:5000
components:
 securitySchemes:
  bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /api/v1/shorts:
    get:
      security:
         - bearerAuth: []
      tags:
        - shorts
      summary: Get all shorts of user
      description: Get all shorts of user
      operationId: getAllShortsOfUser
      parameters:
        - name: userId
          in: query
          schema:
            type: string
            example: 646ef3637251a0220e25132a
      responses:
        '200':
          description: ''
  /api/v1/shorts/{id}:
    get:
      security:
         - bearerAuth: []
      tags:
        - shorts
      summary: Get short
      description: Get short
      operationId: getShort
      parameters:
        - name: a
          in: query
          schema:
            type: string
            example: ''
      responses:
        '200':
          description: ''
  /api/v1/shorts/upload:
    post:
      security:
         - bearerAuth: []
      tags:
        - shorts
      summary: Upload short
      description: Upload short
      operationId: uploadShort
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                duration:
                  type: string
                  example: '1234'
                genres[]:
                  type: string
                  example: '["{id}"]'
                short:
                  type: string
                  format: binary
      responses:
        '200':
          description: ''
...
  /api/v1/shorts/disupvote/{id}:
    post:
      security:
         - bearerAuth: []
      tags:
        - shorts
      summary: Disupvote
      description: Disupvote
      operationId: disupvote
      responses:
        '200':
          description: ''
  /api/v1/shorts/downvote/{id}:
    post:
      security:
         - bearerAuth: []
      tags:
        - shorts
      summary: Downvote
      description: Downvote
      operationId: downvote
      responses:
        '200':
          description: ''
  /api/v1/shorts/disdownvote/{id}:
    post:
      security:
         - bearerAuth: []
      tags:
        - shorts
      summary: Disdownvote
      description: Disdownvote
      operationId: disdownvote
      responses:
        '200':
          description: ''
...
  /api/v1/shorts/scrutinize/{id}:
    post:
      tags:
        - shorts
      summary: Scrutinize
      description: Scrutinize
      operationId: scrutinize
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  example: refuse
            example:
              action: refuse
      responses:
        '200':
          description: ''
...